dav-controller/requirements.txt
fastapi==0.96.0
jinja2==3.1.2
pymongo==4.3.3
python-multipart==0.0.6 # required by fastapi to serve/upload files
qrcode[pil]==7.4.2
structlog==23.1.0
uvicorn[standard]==0.22.0
python-socketio==5.8.0 # required to run websockets
canonicaljson==2.0.0 # used to provide unique consistent user identifiers
pyyaml==6.0.1
cachetools==5.3.3
requests==2.32.2